# perf: improve vi.mock performance