--- a/packages/vitest/src/utils/source-map.ts
+++ b/packages/vitest/src/utils/source-map.ts
@@ -27,6 +27,52 @@ function extractLocation(urlLike: string) {
   return [parts[1], parts[2] || undefined, parts[3] || undefined]
 }
 
+// Based on https://github.com/stacktracejs/error-stack-parser
+// Credit to stacktracejs
+export function parseSingleStack(raw: string): ParsedStack | null {
+  let line = raw.trim()
+
+  if (line.includes('(eval '))
+    line = line.replace(/eval code/g, 'eval').replace(/(\(eval at [^()]*)|(,.*$)/g, '')
+
+  let sanitizedLine = line
+    .replace(/^\s+/, '')
+    .replace(/\(eval code/g, '(')
+    .replace(/^.*?\s+/, '')
+
+  // capture and preserve the parenthesized location "(/foo/my bar.js:12:87)" in
+  // case it has spaces in it, as the string is split on \s+ later on
+  const location = sanitizedLine.match(/ (\(.+\)$)/)
+
+  // remove the parenthesized location from the line, if it was matched
+  sanitizedLine = location ? sanitizedLine.replace(location[0], '') : sanitizedLine
+
+  // if a location was matched, pass it to extractLocation() otherwise pass all sanitizedLine
+  // because this line doesn't have function name
+  const [url, lineNumber, columnNumber] = extractLocation(location ? location[1] : sanitizedLine)
+  let method = (location && sanitizedLine) || ''
+  let file = url && ['eval', '<anonymous>'].includes(url) ? undefined : url
+
+  if (!file || !lineNumber || !columnNumber)
+    return null
+
+  if (method.startsWith('async '))
+    method = method.slice(6)
+
+  if (file.startsWith('file://'))
+    file = file.slice(7)
+
+  // normalize Windows path (\ -> /)
+  file = resolve(file)
+
+  return {
+    method,
+    file,
+    line: parseInt(lineNumber),
+    column: parseInt(columnNumber),
+  }
+}
+
 export function parseStacktrace(e: ErrorWithDiff, full = false): ParsedStack[] {
   if (!e)
     return []
@@ -37,53 +83,13 @@ export function parseStacktrace(e: ErrorWithDiff, full = false): ParsedStack[] {
   const stackStr = e.stack || e.stackStr || ''
   const stackFrames = stackStr
     .split('\n')
-    // Based on https://github.com/stacktracejs/error-stack-parser
-    // Credit to stacktracejs
     .map((raw): ParsedStack | null => {
-      let line = raw.trim()
+      const stack = parseSingleStack(raw)
 
-      if (line.includes('(eval '))
-        line = line.replace(/eval code/g, 'eval').replace(/(\(eval at [^()]*)|(,.*$)/g, '')
-
-      let sanitizedLine = line
-        .replace(/^\s+/, '')
-        .replace(/\(eval code/g, '(')
-        .replace(/^.*?\s+/, '')
-
-      // capture and preserve the parenthesized location "(/foo/my bar.js:12:87)" in
-      // case it has spaces in it, as the string is split on \s+ later on
-      const location = sanitizedLine.match(/ (\(.+\)$)/)
-
-      // remove the parenthesized location from the line, if it was matched
-      sanitizedLine = location ? sanitizedLine.replace(location[0], '') : sanitizedLine
-
-      // if a location was matched, pass it to extractLocation() otherwise pass all sanitizedLine
-      // because this line doesn't have function name
-      const [url, lineNumber, columnNumber] = extractLocation(location ? location[1] : sanitizedLine)
-      let method = (location && sanitizedLine) || ''
-      let file = url && ['eval', '<anonymous>'].includes(url) ? undefined : url
-
-      if (!file || !lineNumber || !columnNumber)
+      if (!stack || (!full && stackIgnorePatterns.some(p => stack.file.includes(p))))
         return null
 
-      if (method.startsWith('async '))
-        method = method.slice(6)
-
-      if (file.startsWith('file://'))
-        file = file.slice(7)
-
-      // normalize Windows path (\ -> /)
-      file = resolve(file)
-
-      if (!full && stackIgnorePatterns.some(p => file && file.includes(p)))
-        return null
-
-      return {
-        method,
-        file,
-        line: parseInt(lineNumber),
-        column: parseInt(columnNumber),
-      }
+      return stack
     })
     .filter(notNullish)