With this change:

```
   Duration  8.67s (transform 1.68s, setup 580ms, collect 1.59s, tests 5.41s)
```

Before the change:

```
   Duration  16.21s (transform 1.70s, setup 29.17s, collect 1.64s, tests 5.34s)
```

We had a single `vi.mock` in `setupFiles`, that was called before each test, but was needed only in one test.

`this.getImporter` took 250-300ms, now it takes ~13ms. This fixes regression, when we added `support-source-map`.